+++ mitol-django-common/mitol/common/pytest_utils.py
@@ -1,10 +1,12 @@
-"""Testing utils"""
+"""Pytest testing utils"""
 import abc
 import json
 import traceback
 from contextlib import contextmanager
 
 import pytest
+
+from mitol.common.envs import generate_app_json
 
 
 def any_instance_of(*cls):
@@ -61,3 +63,22 @@ class MockResponse:
     def json(self):
         """ Return json content"""
         return json.loads(self.content)
+
+
+def test_app_json_modified():  # pragma: no cover
+    """
+    Pytest test that verifies app.json is up-to-date
+
+    To use this, you should import this into a test file somewhere in your project:
+
+    from mitol.common.pytest_utils import test_app_json_modified
+    """
+    with open("app.json") as app_json_file:
+        app_json = json.load(app_json_file)
+
+    generated_app_json = generate_app_json()
+
+    # pytest will print the difference
+    assert json.dumps(app_json, sort_keys=True, indent=2) == json.dumps(
+        generated_app_json, sort_keys=True, indent=2
+    ), "Generated app.json does not match the app.json file. Please use the 'generate_app_json' management command to update app.json"